HUE-4250 [editor] Add configurable timeout for autocompletion API calls
This adds a timeout with default value of 3 seconds when fetching tables, columns, fields, samples and hdfs paths from the API for autocompletion. It will still suggest cached values, from browsing the assist etc. Note that this is only for autocompletion, the assist is not affected by this change.

When the timeout value is set to 0 the autocompleter will never call the API.