HUE-4018 [editor] Error when canceling a query that is slow being submitted
Check for empty handle in case the operation was not even submitted.
Abort ajax of blocking submit operation.

If the query was running, we cancel like before by sending a cancel of the statement.
If the query was not even submitted yet but blocks, we abort the submit.
If the query is fetching data but blocks, we abort the fetch.