HUE-4320 [editor] New autocompleter should make suggestions based on types in functions
It now makes the right type completions based on return types and arguments for UDFs. For CAST it'll figure out the correct return type but not for other functions like T = x(INT, T) where it will suggest anything (T).

Arrays, maps and structs are always suggested as you could do cos(structCol.structCol.number), for this type of UDF arguments, like in explode(), it will only suggest functions or columns that are structs, arrays and maps.

For the known functions it will only make completions for argument lengths that are valid i.e. sin(|) but not sin(1,2,3,|). If the function is unknown it will make completions like before.