HUE-4414 [editor] Allow to send queries even after closing the session
Clicking on Close session goes to a dead end. Now we still allow the automatically
created new session to be used.

Ideally we will update the UI instead of not displaying the new session at some point.

Note that the session can be closed already if HiveServer2 or Hue were restarted before
clicking on the Close button.

Restart Button works well.