HUE-4431 [editor] Batch mode does not always update the query
We need to convert the current query to the history notebook and continue with it.
If not, we use the content of the last notebook from the last time it was saved by the user.

Changing the flow by creating the history, then executing it, then updating it (before was
just executing it, then creating the history).