Assure make does not fail because the wrong version of python is used to configure
In this early stage the Makefile.vars file is executed without knowing exactly which version "python" refers to. This causes failures if your default "python" is python 3, even printing the version, etc. I have changed all the python calls to be 2/3 compatible (for instance adding parentheses to the arguments to the print call, that works equally well in python 2 and 3).