HUE-4725 [editor] The autocompleter should handle Impala complex types in the table list
This is a rather large refactoring of how tables and databases are represented in the suggestions. For Impala it's impossible to know, based on just the statement, whether a reference in the table list after FROM is for a database and table or table and column. Because of this It's now an array of identifiers in the suggestions instead of specific database or table.

This also improves completions throughout