HUE-4938 [editor] Fix issue where the map crashes the browser for large coordinate values
This limits the longitude and latitude within the allowed bounds, when coordinates are outside the bounds the max and min values will be used for longitude and latitude instead. This also improves the automatic selection of chart X and Y when there is only 1 or 2 columns in the result. It also takes care of an issue where the map markers were not reset after multiple queries.