HUE-5002 [editor] Improve nav tag edit mode layout
This puts the edit icon next to the last tag, the save and cancel actions are still in upper right for now as it's not trivial to have them next to the last tag with Selectize active. The save and close are now always visible and slightly larger.

(cherry picked from commit 8b558b967d735ba2c4374ac13e0f2bb87005807d)