HUE-5048 [frontend] Remove RequireJS dependencies
RequireJS has been causing some pain with the asynchronous loading, especially in MS Edge. At this point we're not loading any modules on demand so there's no major drawback in dropping the asynchronous part. We're still keeping the modules and until we find a better alternative we need to specify all the required modules in the script tags of the main app page.