HUE-5042 [core.backend] Unable to kill jobs after Resource Manager failover
This is a direct change to origin/cdh5-3.9.0_5.7.x_dev. It is not committed on Master.

In 5.7.x, The issue is RM switchover taking time. Basically when Active RM goes down, Standby RM thinks it is Standby and not promoted to Active state. In this change we have implmented 're-try' logic.