HUE-5325 [aws] Enable S3 browser on only when keys are there
Permission is in FileBrowser, not the AWS lib.
Do not add S3 permissions to the default group.
Fix the logic to block S3 access via File Browser if the user does not have
the permissions or is not a Hue admin.

(cherry picked from commit 7c31aa116e6523277fa95c94ea99aec7650c6659)