HUE-6180 [beeswax] Extend _verify_query_state in beeswax tests to allow for multiple states (#514)
In test_basic_flow beeswax test, we found a case where the query whose
state was being verified had actually completed before the test got the
chance to check the state, causing what was effectively a race condition
between the query engine and the test framework. This change allows for
the query to be in multiple states, which will still pass the test even
if the query finishes "too quickly".