HUE-7713 [search] Support handling indexing of documents with bad fields
This is only for Solr 6+ and creation created via Hue or having a 'tolerant' updateprocessor chain available.
It will not fail the ingest until a maximum of 100 records fails to be indexed. The reason of failure (e.g.
why one of more fields are not compatible with the type) for the records will be displayed.

(cherry picked from commit 7ce795458623824496b1a214443b8616ddb92b92)