HUE-9028 [editor] Add support for select to execute multiple in notebook 2
When selecting multiple statements it will execute all of them in order without stopping as we now store the result per statement. It's also possible to execute multiple chains in parallel by selecting non-intersecting chains.

If a previous selected execution chain is intersecting a new execution it will cancel the previous chain.