HUE-9037 [core] Add .cloudera.com to trusted host for CSRF token
When using https, Django will look at "HTTP_REFERER" request header to
see if it matches the current request's host when using unsafe (e.g.
POST) requests. When using Hue via Knox, the Knox host is the referrer
and if it is on a different host than the Hue host then it does not
match. To handle this, HUE-8750 adds the Knox host to
Django's CSRF_TRUSTED_ORIGINS and the Knox host is typically configured
via Cloudera Manager. Unfortunately, in Cloud scenario, the external
Knox host name may be different then the internal Knox host name and the
check fails yet again. To handle this, we're adding a new setting that
adds '.cloudera.com' as a trusted host.

Change-Id: Ida09bce0add7cc8a6ab89397fd3a9b580f7bee5b