HUE-9073 [fb] Only enable IDBroker when on cloud.
We currently enable S3 when IDBroker is configured with S3:
https://github.com/cloudera/hue/blob/6db70177fc1c6955340da4d7475cdda5c44c2aef/desktop/libs/aws/src/aws/conf.py#L248,
https://github.com/cloudera/hue/blob/6db70177fc1c6955340da4d7475cdda5c44c2aef/desktop/core/src/desktop/lib/idbroker/conf.py#L60
However, I'm noticing on that IDBroker is always configured with all 3
cloud providers.
Furthermore, the CM cloud connector is not being used in CDP so CM is
currently not configuring Hue with anything. For now only enable cloud
storage is it's specifically configured or if on cloud and IDBroker
enabled.

(cherry picked from commit 63bcf034935dc345fc9d4b863c9416bc40358045)

Change-Id: Ib207948dcceee5da948c9826e2b65833bfda807a