HUE-9085 [fb] Improve S3 bucket 301 & 400 error message
When retrieving a bucket a 301 error signify that the bucket exists, but
it is in a different region than the one we are connected to. A 400
error can mean a few things, one of which is that we're connected using
V2 and we're trying to access a region that requires V4 signature. V4
signature requires specifying a host or region when connecting.