HUE-9089 [hive] Upgrade Thrift protocol to V11 to support LLAP
As part of task support, LLAP requires HIVE_CLI_SERVICE_PROTOCOL_V8 and
changes the behavior of hasResultSet in TOperationHandle as it can be
updated in TGetOperationStatusResp. This is specific to LLAP as TEZ
keeps the old behavior of settings hasResultSet in
TExecuteStatementResp.
In further commits, we might want to also use TProgressUpdateResp that
came in as part of HIVE_CLI_SERVICE_PROTOCOL_V10 to track query progress
as an alternative to looking at Yarn for job progress.