HUE-9221 [tools] Add a tool that automatically adds review and commit urls to Jira comments
This tool uses rbt and git cli as well as the Jira API to sync missing details about commits and reviews in the Jira comments.

To use it, first run 'npm install' in tools/jira/ then try it out in a dry run:

  node tools/jira/hueJira.js -r <rbtUser> -j <jiraUser> -p <jiraPassword> -d

Without -d it will also add the missing comments to the jiras automatically.