Hue Aquascan CVEs (#1095)
"This commit does not contain secrets"

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
 - Adding Django-1.11.29

* HUE-8905 [core] Apply HUE-8772 to Django-1.11.22 for fixing 'user is missing in mako context'

(cherry picked from commit cf717de5cd857701535c754cb242a97f08bd9214)

* HUE-8905 [core] Apply HUE-8836 to Django-1.11.22 for HTTP_X_FORWARDED_HOST contains multiple hosts

(cherry picked from commit 26e5f7a096fd1b05590ad5a33e43b4e252dc8c57)

* Revert "HUE-8905 [core] Apply HUE-8772 to Django-1.11.22 for fixing 'user is missing in mako context'"

This reverts commit cf717de5cd857701535c754cb242a97f08bd9214.

(cherry picked from commit 9014bfa9b9c4773a4186f3f64fc8e2616ad7eeee)

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Removing Django-1.11.22

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding cryptography-2.9
- Removing cryptography-2.1.4

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding MarkupSafe-1.1.1
- Removing MarkupSafe-0.9.3

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding PyYAML-5.3.1
- Removing PyYAML-3.12

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding requests-2.23.0
- Removing requests-2.18.4

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding urllib3-1.22
- Removing urllib3-1.25.8

* HUE-5095 [backend] Python requests library should put port information in log message

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding requests-kerberos-0.12.0
- Removing requests-kerberos-0.6.1

* HUE-7127 [backend] Fix for Python requests Kerberos/GSSAPI authentication library fails to authenticate kerberos requests to the same destination

current python request Kerberos library fails generating the GSSAPI authentication token with kerberos for the host which run multiple kerberised

Testing Done:
- Manual testing
  - using multiple load generator scripts
- Tested on different cluster with Python 2.6 and Python 2.7

* HUE-8202 [jb] Fix mutual authentication failed with Isilon (#675)

https://review.cloudera.org/r/12820/

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding tablib-0.14.0
- Removing tablib-0.10.0

* HUE-2523 [core] Remove xlwt and xlrd from tablib

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding MarkupPy-1.14

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding openpyxl-2.6.4
- Removing openpyxl-2.3.0-b2
- Removing openpyxl-2.5.3

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding odfpy-1.4.1

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding pysaml2-4.9.0
- Removing pysaml2-4.4.0

* HUE-3102 [libsaml] Add support for private key passwords

This is also tracked by IdentityPython/pysaml2#278.

This patch adds support for SAML certificates that are protected
with a password. The way it does so is with a bit of trickiness,
due to the fact that `xmlsec1`, which is an external program that
pysaml2 uses to sign the XML requests, which does not have great
support for password protected certificates. It either supports
passing in the password on the command line (which is not safe
since someone else on the machine could see the password), or
through an interactive prompt.

The proper way to fix this would be to update pysaml2 to use
another xmlsec library, but implementing that may take some
time. In the short/medium term, this patch implements this
instead by decrypting the certificate in memory, and passing
this decrypted certificate to xmlsec1 through a named pipe.
This protects us from the decrypted certificate ever hitting
the disk.

Unfortunately, this solution is only portable to
POSIX-compatible platforms. That is fine for Hue, but it
probably means we cannot push this patch to the upstream
pysaml2 repository. This patch will tied us over until
the upstream project switches to a better xmlsec library.

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding chardet-3.0.4

* HUE-9225 [core] Upgrade certain third party python libraries that has identified vulnerabilities
- Adding pycryptodomex-3.9.7
- Removing pycrypto-2.6.1
- Removing pycryptodomex-3.4.7

Co-authored-by: Ying Chen <yingchen@cloudera.com>
(cherry picked from commit b99879b7a0faf5dc1492ad9fa102ec47510e6b02)
Change-Id: Ia94f169a087ac6329697a988be1ff9d40742c61d