Increase username length to 150 characters
On https://github.com/cloudera/hue/pull/1104 (HUE-8020), username length
was increased up to 150 characters. Anyway, some backends (at least,
LDAP) are still validating that username is up to 30 characters.

This increases the validation lenght to match with the size of the field
on the DB and the max input lenght on the authentication form.