[oozie] Fix usage of iterkeys() and zip() in list_oozie_info template (#1270) (#1271)
The function iterkeys() is not available in Python 3, but there seems
to be no real need for an iterator anyway. The zip() call returns an
interator in Python 3, not a list, so the compromise of wrapping
list(zip()) should suffice for this use case.

Co-authored-by: Romain Rigaux <romain.rigaux@gmail.com>
(cherry picked from commit d6c57a4d4010fc95fb9dac8925b0d4262021e811)