[notebook] The default list of interpreter is not cached properly
Here we get the proper list of default interpreters.
Before this list depended on the interpreters of the first user landing
on the server.

And now the default list is then filtered properly with the permissions
of the loged in user.

Note that `_default_interpreters(user)` was auto caching the values
which made successsive call lose the extra interpreters.

This was also a slow operation (happening quite often).

All this logic is gone and simpler when using the new connectors.