[jobbrowser] Add Hive kill query via Notebook call
Call is done as the logged-in user making the call and will
go throught the user existing HiveServer2 session or create
a new session.

$.post("/jobbrowser/api/job/action/queries-hive/kill", {
  "operation": JSON.stringify({'action': 'kill'}),
  "interface": JSON.stringify('queries-hive'),
  "app_ids": JSON.stringify(['d94d2fb4815a05c4:b1ccec1500000000'])
}, function(data) {
  console.log(JSON.stringify(data));
});

Using the notebook API is a good skeleton example for when we could
have spefici actions like kill for other SQL dialects.

Note: not compatible with connectors.