[catalog] Move the data catalog code to typescript and switch form jQuery deferrals to Promises
This is a rather big change but it allows us to use proper promises when working with the data catalog and enables us to switch over to Axios easier.

This also takes care of a couple of "undefined exception" bugs related to the catalog and it's now 100% typed.

(cherry picked from commit 368afbe6c92c3d8a130bf6bdf754393450cff4b8)