[editor] Separate the editor v2 from the notebook
This introduces a new webpack entrypoint for the editor v2 and it moves the feature flag test to the backend. If the flag is enabled it will return editor2.mako, for the editor and the old notebook.mako for the notebook.

This gives us full control of the editor v2 page layout without interfering with editor v1 and the notebook functionality.

Notebook will keep using editor v1.

(cherry picked from commit 0a9a8a7c205cf179e93058fecadc6d8a50ec8985)