[editor] Do not auto load sample when creating a brand new editor
All the links (e.g. from menu) will auto load the samples, but we should
still allow a blank editor.

This is only affecting when auto load samples is true (i.e. on
demo.gethue.com).

Did not see a simpler way to do. Could have gone the other way around
and changed the URLs of the editor with the new parameters in
get_config() but this feels shorter and less intrusive.