[core] Add request as positional arg for RemoteUserDjangoBackend
This patch is needed for Django 2.1+ since the
django.contrib.auth.authenticate() function requires 'request' as
positional argument. In #2004 we partially addressed the issue, but not
completely.

This commit is present only in Django 1.11's and 2.0's branches:
https://github.com/django/django/commit/3008f30f194af386c354416be4c483f0f6b15f33

There is a warning msg stating that 'request' should be moved as first
positional argument.