[frontend] Use relative imports in exported NPM package sources
This allows consumers of the NPM lib to use any code inside Hue given that they use the required webpack plugins and babel loaders.

By using relative imports there's no need to set it as an additional module root, limiting transpile to the actual used modules as opposed to the entire Hue source.