[frontend] Merge Executable into SqlExecutable
This takes care of an issue with inheritance of abstract classes in web components. In reality we only have one type of executable so there's no actual need for the abstraction. When and if we need more executables it's better to use an interface.

(cherry picked from commit 4e395545a2abb683105910a9fb737e99edbd3068)