[importer] Prepare import from a Kafka Topic into a Table
This is part of a series of commits, with tests following-up next
as well as more polishing of the current inputs of step 1.

The goal of this PR is to make the POCs easier to demo and understand.

All is flagged of by default and is leveraing the previous design and
its past implementation to keep is as simple as possible and easy to
evolve.