[sparksql] Enhance SparkSQL support via Apache Livy for Hue (#2794)
* [sparksql] Enhance SparkSQL support via Apache Livy for Hue

- Reuse Livy session per user similarly to SQLAlchemy
- Add support for /autocomplete, /get_sample_data and other APIs
- Add few UTs and update existing ones

NOTE: In the start, 2 sessions are being created, one for /autocomplete and other for /create_session. This is because /autocomplete call is always first in the call stack from the UI, plus this will help if user just want to access other tabs like Table Browser where we don't do /create_session call. But since we are reusing the session per user, eventually we only use one of the sessions afterwards for every user activity regarding SparkSQL.