[autocomplete] make sql autocomplete and syntax parsers pluggable (#2861)
* [autocomplete] fixed failure when generating new parser

* [autocomplete] separate module imports from sqlParserRepository and generate new file on make apps

A new file called parserModules.ts which contains the import statements for all available parsers is now generated when running make apps. The generation of the import statements has been removed from generateParsers.js and is now handled by the generateParserModuleImports.js.

A key difference is that the import statements are generated based on the actual available parser files rather than the available Jison parser definition files. This means that a new parser can be created and maintained in a separate project and by adding its generated parser files to the “parser folder” in downstream Hue (desktop/core/src/desktop/js/parse/sql) it will be automatically picked up by the sqlParserRepository.ts

* [ci] add call to generateParserModuleImports

Co-authored-by: Björn Alm <balm@cloudera.com>