[knox] Disabled the MultipleProxyMiddleware for py2 (#3021)
The middleware is needed only on py3 and is interfering when
authenticating with knox. We are stripping out the gateway host's
header information and that is why it is failing the match
with proxy_hosts config.

(cherry picked from commit f4d0ebcbe063fbf3a3d9a5c349dc90152babce16)