[frontend] Add support for parser tests next to the jison definition
This allows us to move the test cases for the autocomplete parser next to the corresponding jison files to make development of parsers easier and to reduce duplication of test cases.

Before this change the tests were specified for each parser even for parsers that were referencing the same jison file. For changes to a specific jison file the tests would have to be updated in multiple places. Now the exact same tests will be ran for each parser that references the same jison file greatly reducing duplication.

This commit also move the alter table tests for the generic parser into the new setup.