[eventlet] Swapped eventlet from download to local install
The goal is to patch eventlet for cipher list fix. The fix is available
upstream (https://github.com/eventlet/eventlet/commit/7aa50ad5) and also
in the latest eventlet release but it is not compatible with our current
gunicorn and not even the latest gunicorn. So, we will patch eventlet
with the single commit from upstream.