CDPD-43652: [fb] Prevent decoding of already decoded path in the File Browser (#3087)
Currently decoding happens at least twice in views.py when requesting to view a folder, this causes an issue when files are stored with actual encoded values in the name.

An example is if a folder exists in hdfs with the name 'some%20folder', the UI would encode it and request 'some%2520folder' (% becomes %25), which if decoded once would be correct, 'some%20folder', but if it's done twice the resulting name would be 'some folder' with a space instead of %20 which is incorrect.

We can safely assume that each path contains a '/' char, so we can use that to check if a path has already been decoded. An encoded path would never contain a '/' as it would be %2F.

(cherry picked from commit 57eabed7fc8aec0cb1b36d151db2e7d7f9090242)
(cherry picked from commit e85730b891f7933713954cbda5ffe89f8693286d)
Change-Id: Ib38090bb2b57e7dc5b014fc27386079ee7613431