[core] Remove the is_embeddable flag throughout
The is_embeddable flag was introduced in Hue 3 as a feature flag for the new Hue 4 layout. The name "embeddable" comes from the switch to a single-page app, where an embeddable page could be embedded in a container element of the page and if is_embeddable is false it'll render as the complete Hue page.

Since the release of Hue 4 the flag is always true except for some cases where a user by mistake would end up on a Hue 3 page, for instance via an old bookmark. We haven't actually removed a lot of those pages and there could be Hue 3 URLs that are still active, right after the release of Hue 4 most Hue 3 pages worked, but that's definitely not the case today. Now the is_embeddable flag is causing a lot of confusion and there's plenty of dead code around, removing the flag is the first step in getting rid of the old pages and layout helpers like commonheader and commonfooter.