[backend/file-browser] Force cache validation for file download
In the files-download, we were always setting the 'Last-Modified' header
for the downloaded files. Since, we were not providing the expiration,
the browser was not being forced to verify for a new versions and are
free to use their own heuristics. See the Caching RFC for more details
https://www.rfc-editor.org/rfc/rfc7234#section-4.2.2

This commit adds the 'Cache-Control: no-cache' header to force browsers
to validate the modified timestamp of the file before reusing the file
from cache.

Steps to reproduce (not reproducible every time):
1. Upload a file using the file browser.
2. Leave it for a few hours. Waiting is important as the heuristics for
    default expiry is dependent on the age of the file itself.
3. Open the JavaScript console.
4. Download the file. It should come with status 200
5. Quickly delete the file and upload another one with the same name.
6. Download the file with same name multiple times, while monitoring the
   response status code.
4. Some of the downloads will come from the cached version.

Steps to verify:
1. Follow the steps 1-3 above
2. The browser should use the cache version only as long as the file
   is not modified on the server.

(cherry picked from commit 53257779fc36025804eb4abc72b8cab1fe7df9b8)