Hue ID Broker HA improvement
Previous commit for helped to address the issue if there were multiple id brokers in the core-site.  But with the new information it seems like the information which is added automatically to the core-site was missing the ending '/' which is causing the same issue,  this fix would address that.

Jira: CDPD-50058