CDPD-50058: Hue ID Broker HA improvement (#3211)
Previous commit for helped to address the issue if there were multiple id brokers in the core-site.  But with the new information it seems like the information which is added automatically to the core-site was missing the ending '/' which is causing the same issue,  this fix would address that.

(cherry picked from commit 7cd42462ec81295c30fde87e2e8ae0b8cdda571b)
Change-Id: I27dc7346827112456d005976dc0013ae104cdf10