[frontend] Add support for showing versions and warehouse name in the left sidebar footer
This will show the warehouse name, image version and hue version if the corresponding attributes are set in the get_config API response ("vw_name", 'img_version" and "hue_version")

When any text is too long it'll hide the overflow and show an ellipsis, the complete value is shown on hover via the title attribute. The image version will take precedence over the hue version when both are set, the hue version will then be shown on hover.