[frontend] Replace ENABLE_NOTEBOOK_2 with ENABLE_HUE_5
The scope of Hue 5 has grown beyond editor/notebook v2 and instead of introducing additional feature flags I've renamed the old one and moved it up in the desktop section of the config.

The ENABLE_HUE_5 flag can be used for any features that we intend to deliver in Hue 5 only.