DWX-15497: Add CDWH-specific NOTICE files to Docker images
as required by Legal.

This commit is based on cherry-picking 799f19f as it was not merged to master.
Adding it there as well now to ease this activity for the future.

The NOTICE files are generated by FOSSA by scanning the Docker images
themselves (using images from a previous run). Contents for the NOTICE
files was retrieved from
https://github.infra.cloudera.com/Kitchen/Fossa-Notice-Reports/tree/master/CDP-PVC/2023.0.15.2,
and placed into the container-build subdirectories of their respective
container configurations.

Change-Id: Ib1f656a7556cf3a9a9e933ad974ee40d1f3b65bf