[jb] Add status filter to the tasks listing of an Oozie schedule in the Job Browser
The implementation of this was left todo many years ago, back when we made the switch from the dedicated Oozie status pages to the Job Browser in Hue 4.

Given that the Job Browser pages are generic in the sense that the same code is used to display many types of "jobs" I've deliberately kept this change as non-intrusive as possible. Existing patterns, variables and values have been re-used even if there's quite some room for improvement and the filter will only appear for the tasks listing in the schedules tab.

(cherry picked from commit d193a45a1020b2a11dc1556da1b429bb5da71a12)