[compute] set default user/pass for hive from config
Our hive authentication relies on the use of doAs parameter. The user
with proxy_privilege (typically system user with enhanced privilege
such as `hive`) is used as auth_username is same for all users. The
currently logged-in user's username is passed in the
`hive.server2.proxy.user` field while making the connection. Since, we
were not setting the default auth_username and password for computes,
it was relying on using the currently logged-in user and was causing
permission issues.

Change-Id: I5cf408462c3a34ccfd44632e400763b90f10bba2