[editor] Remove mako variables for editor id and notebooks as well as the old notebooks listing page
In this commit I'm removing the editor_id and notebooks that gets set from the backend when opening the editor. The editor ID is now taken from the url parameter and notebooks isn't actually used anymore.

As part of this I'm also removing the notebooks listing page which should be considered dead code as it's no longer reachable from the UI or via URL.

This is part of the effort of removing inline script and our goal of moving towards CSR.