[impala] improved query timeout error logging
we were not clearly catching the timeout error from impala query
submission. the error response was getting passed to the upper layers
that errored out when trying to get json out of it. It lead to a cryptic
error on the frontend as well. Now we are generating a clearer log
message with the timeout values and config name. The frontend also
shows a timeout error message.

Change-Id: I19fbf49fd0fccc21f8f1f2258695902f73406856