CDPD-72775: [doc2] Fix issue where a copy of a shared document in user home becomes unreachable
This takes care of an issue where a copy of a shared document ends up in the original owners home directory and becomes unreachable by the user making the copy. With this change these copied files are now put in the home directory of the user that makes the copy instead. If the file to be copied exists in a shared folder the copy will end up in that same shared folder, a user being able to copy such a document implies that the user is also able to browse that directory.

(cherry picked from commit a2c3090cbae4e4875ffa601c72f885b59f5bce37)
(cherry picked from commit 14043bada62a26d88d888ef23dcdaaf093ef1bea)
Change-Id: I4f9f54588a2736234aa37664761b8175e1a44712